% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensustools.R
\name{census_table_timeseries}
\alias{census_table_timeseries}
\title{Census Table Time Series}
\usage{
census_table_timeseries(census_table, geography, state, study_areas, years)
}
\arguments{
\item{census_table}{character strings of census table id}

\item{geography}{character string of geography that the data is available at (tract, county, zcta, etc...)}

\item{state}{character string of state study areas are in (without this, downloads all counties in US)}

\item{study_areas}{vector of character strings of areas that the table allows for}

\item{years}{vector of numeric type for each year desired}
}
\value{
Returns: tibble dataframe in wide format
#'
}
\description{
This function returns a dataframe of specified census table within designated areas and years
}
\section{Features to Implement}{

Todo list of things to add

\itemize{
  \item Grab geographies from multiple states/counties
  \item Grab variables from different tables
  \item Return object in long format
}
}

\examples{
census_table_timeseries("B25014", "county", "CA", ("Los Angeles", "Orange"), (2019,2018,2017))

}
