% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_by_hispanic_rename_vars.R
\docType{data}
\name{race_by_hispanic_rename_vars}
\alias{race_by_hispanic_rename_vars}
\title{Rename column names for Race by Hispanic Census Table}
\format{
A data frame with 23 rows and 2 variables:
\describe{
  \item{column_name}{census defined lengthy variable names}
  \item{desired_name}{self defined short variable names}
  ...
}
}
\usage{
race_by_hispanic_rename_vars
}
\description{
A dataset containing lengthy named census variables and alternative short
variable names used to rename variables
}
\keyword{datasets}
